@@ -1,12 +1,18 @@
-import Vue from 'vue'
-import Router from 'vue-router'
-import LoginForm from '../components/LoginForm.vue'
-import DashBoard from '../views/DashBoard.vue'
-// import ToolBar from '../components/ToolBar.vue';
 
-Vue.use(Router)
+/**
+ * router/index.ts
+ *
+ * Automatic routes for `./src/pages/*.vue`
+ */
 
-export default new Router({
+// Composables
+import { createRouter, createWebHistory } from 'vue-router/auto'
+import { routes } from 'vue-router/auto-routes'
+import LoginForm from '../pages/LoginForm.vue'
+import DashBoard from '../pages/DashBoard.vue'
+
+const router = createRouter({
+  history: createWebHistory(import.meta.env.BASE_URL),
   routes: [
     {
       path: '/',
@@ -19,5 +25,26 @@ export default new Router({
       component: DashBoard,
       props: {}
     }
-  ]
+  ],
 })
+
+// Workaround for https://github.com/vitejs/vite/issues/11804
+router.onError((err, to) => {
+  if (err?.message?.includes?.('Failed to fetch dynamically imported module')) {
+    if (!localStorage.getItem('vuetify:dynamic-reload')) {
+      console.log('Reloading page to fix dynamic import error')
+      localStorage.setItem('vuetify:dynamic-reload', 'true')
+      location.assign(to.fullPath)
+    } else {
+      console.error('Dynamic import error, reloading page did not fix it', err)
+    }
+  } else {
+    console.error(err)
+  }
+})
+
+router.isReady().then(() => {
+  localStorage.removeItem('vuetify:dynamic-reload')
+})
+
+export default router
